backend/requirements.txt
# Core Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.0

# ML & Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Document Processing & OCR
pytesseract>=0.3.10
Pillow>=10.0.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
pdf2image>=1.16.0
spacy>=3.7.0
httpx>=0.25.0

# Utilities
requests>=2.31.0
aiofiles>=23.2.0

# Report Generation
reportlab>=4.0.0

# AI/LLM
google-generativeai>=0.3.0